+++ app/src/androidTest/java/ch/epfl/reminday/ui/activity/MainActivityInstrumentedTest.kt
@@ -13,7 +13,7 @@ import androidx.test.ext.junit.rules.ActivityScenarioRule
 import ch.epfl.reminday.R
 import ch.epfl.reminday.data.birthday.BirthdayDao
 import ch.epfl.reminday.testutils.UITestUtils
-import ch.epfl.reminday.utils.Mocks
+import ch.epfl.reminday.util.Mocks
 import dagger.hilt.android.testing.HiltAndroidRule
 import dagger.hilt.android.testing.HiltAndroidTest
 import kotlinx.coroutines.runBlocking
