+++ app/src/main/java/ch/epfl/reminday/adapter/BirthdayAdapter.kt
@@ -12,7 +12,7 @@ import ch.epfl.reminday.data.birthday.Birthday
 import ch.epfl.reminday.databinding.FragmentBirthdayListItemBinding
 import ch.epfl.reminday.format.date.DateFormatter
 import ch.epfl.reminday.ui.activity.BirthdaySummaryActivity
-import ch.epfl.reminday.utils.ArgumentNames
+import ch.epfl.reminday.util.constant.ArgumentNames
 import java.util.*
 
 class BirthdayAdapter(
