+++ app/src/main/java/ch/epfl/reminday/util/Extensions.kt
@@ -1,4 +1,4 @@
-package ch.epfl.reminday.utils
+package ch.epfl.reminday.util
 
 import android.text.Editable
